Move CancellableServerCall to a more reasonable location
Principally, make it match its internal location.
