+++ atlasdb-commons/src/main/java/com/palantir/annotations/remoting/CancelableServerCall.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Palantir Technologies, Inc. All rights reserved.
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
  *
  * Licensed under the BSD-3 License (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marathon.util.spring;
+package com.palantir.annotations.remoting;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
